04_AIML_CRASH_COURSE/INTRO_TO_GENERATIVE_AI/02_TEXT_2_SQL/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "openai>=1.65.4",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "pydantic>=2.10.6",
    "sift-stack-py>=0.4.2",
    "streamlit>=1.43.1",
]
